Update template from NextUI to HeroUI
The project template has been updated to use HeroUI instead of NextUI.
This includes changes to the README.md file, such as the title, description,
and links to documentation. The command for creating a new project with
`create-next-app` has also been updated to point to the HeroUI template
repository. Additionally, the footer in the layout component has been
modified to reflect the use of HeroUI.